Rename and relocate shared auth layout components
